+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0]
+
+* add surrogate keys for image variants
+* add vary and cache-control headers for DMS documents
+
 ## [1.0.0]
 
 initial release